+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [7.5.0] 2025-04-30
+
+### Added
+
+- [TD-5800] Add validations for mandatory fields in templates
+
+### Fixed
+
+- [TD-7226] Enhance SSL configuration handling in production
+
 ## [7.4.0] 2025-04-09
 
 ### Changed